refactor(structure): restructure project modules and update build configuration
Why:
- Improve architectural purity by decoupling platform-specific configurations from the shared logic layer, adhering to Clean Architecture principles.
- Optimize web performance and reduce maintenance overhead by standardizing on WasmJS and removing the legacy JS target.
- Enhance build reliability and developer experience by adopting type-safe project accessors to catch dependency errors at compile-time.
- Standardize the project structure with modern Kotlin and Gradle best practices to improve maintainability and internationalization.

What:
- Switch `desktopApp` to the Kotlin JVM plugin and relocate source files from `jvmMain` to `main`.
- Move `compose.desktop` application configuration from the `shared` module to `desktopApp`.
- Remove the JS target from the `webApp` module, focusing on WasmJS.
- Refine the `shared` module by removing platform-specific dependencies and updating Android test configurations.
- Transition to type-safe project accessors (e.g., `projects.shared`) across modules.
- Update `libs.versions.toml` with new library definitions for `kotlin-wrappers` and Compose UI tooling.
- Translate code comments in the web module to English and remove debug print statements.
