+++ build.gradle.kts
@@ -11,6 +11,7 @@ plugins {
     alias(libs.plugins.kotlin.cocoapods) apply false
     alias(libs.plugins.kotlin.serialization) apply false
     alias(libs.plugins.kotlin.parcelize) apply false
+    alias(libs.plugins.kotlin.jvm) apply false
     alias(libs.plugins.buildkonfig) apply false
 }
